+++ annoworkcli/workspace_tag/subcommand.py
@@ -2,8 +2,8 @@ import argparse
 from typing import Optional
 
 import annoworkcli
-import annoworkcli.organization_tag.list_organization_tag
-import annoworkcli.organization_tag.put_organization_tag
+import annoworkcli.workspace_tag.list_organization_tag
+import annoworkcli.workspace_tag.put_organization_tag
 
 
 def parse_args(parser: argparse.ArgumentParser):
@@ -11,8 +11,8 @@ def parse_args(parser: argparse.ArgumentParser):
     subparsers = parser.add_subparsers(dest="subcommand_name")
 
     # サブコマンドの定義
-    annoworkcli.organization_tag.list_organization_tag.add_parser(subparsers)
-    annoworkcli.organization_tag.put_organization_tag.add_parser(subparsers)
+    annoworkcli.workspace_tag.list_organization_tag.add_parser(subparsers)
+    annoworkcli.workspace_tag.put_organization_tag.add_parser(subparsers)
 
 
 def add_parser(subparsers: Optional[argparse._SubParsersAction] = None) -> argparse.ArgumentParser:
